Minor refactor in HTTP tests
