+++ src/test/java/io/vertx/core/http/Http1xMetricsTest.java
@@ -16,11 +16,4 @@ public class Http1xMetricsTest extends HttpMetricsTestBase {
     super(HttpVersion.HTTP_1_1);
   }
 
-  @Override
-  public void setUp() throws Exception {
-    super.setUp();
-    client = vertx.createHttpClient(new HttpClientOptions());
-    server = vertx.createHttpServer(new HttpServerOptions().setPort(DEFAULT_HTTP_PORT).setHost(DEFAULT_HTTP_HOST).setHandle100ContinueAutomatically(true));
-  }
-
 }
